mcp-server-seedream/pyproject.toml
[project]
name = "mcp-server-seedream"
version = "0.1.0"
description = "MCP Server for 火山引擎即梦Seedream 4.0 API"
requires-python = ">=3.10"
dependencies = [
    "fastmcp>=2.0.0",
    "httpx>=0.27.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/mcp_server_seedream"]